project ompiles. Interface communicatiuon facade defined

+++ app/src/main/java/com/example/examscanner/repositories/grader/GraderRepository.java
@@ -28,7 +28,7 @@ public class GraderRepository implements Repository<Grader> {
     @Override
     public Grader get(long id) {
         //TODO - implement comFacade get Grader
-        return new ExamManager(id);
+        return new ExamManager((int) id);
     }
 
     @Override
